requirements.txt
django==5.2.6
django-cors-headers==3.7.0
django-debug-toolbar==5.0.0a0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
django-environ==0.4.4
drf-spectacular==0.28.0
gunicorn==23.0.0
PyJWT==2.9.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
cryptography==41.0.7
pycryptodome==3.19.0
pillow==11.3.0
PyPDF4==1.27.0

django-simple-history==3.8.0

#ia_assistant
django-ai-assistant==0.1.1
fabric==3.2.2

# LangChain and providers for django-ai-assistant
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.1
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.28

# Google Calendar Integration
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.104.0


pyotp==2.9.0

psycopg2-binary==2.9.9


